# Conflicts:
#   code/test/controller.unit.test.js

+++ code/controllers/controller.js
@@ -419,10 +419,10 @@ export const getTransactionsByUser = async (req, res) => {
             if(!authorized) return res.status(401).json({error: cause})
             
 
-            filters = {};                        
+            filters = {username : req.params.username};                        
         } else {
-            // regular user authentication            
-            const {authorized, cause} = verifyAuthUser(req, res, req.params.user);            
+            // regular user authentication                        
+            const {authorized, cause} = verifyAuthUser(req, res, req.params.username);            
             if(!authorized) return res.status(401).json({error: cause})
 
             const amountFilter = handleAmountFilterParams(req);
@@ -432,6 +432,7 @@ export const getTransactionsByUser = async (req, res) => {
 
             filters["$and"].push(amountFilter);
             filters["$and"].push(dateFilter);
+            filters["$and"].push({username : req.params.username});
         }                  
 
         const {username} = req.params;
@@ -530,7 +531,8 @@ export const getTransactionsByUserByCategory = async (req, res) => {
             [
                 {
                     $match : {
-                        type
+                        type,
+                        username
                     }    
                 },
                 {
@@ -785,7 +787,7 @@ export const deleteTransaction = async (req, res) => {
         }
 
         // body is not complete
-        if (!id) return res.status(400).json({ error: 'body does not contain all the necessary attributes' })
+        if (!id || id === '') return res.status(400).json({ error: 'body does not contain all the necessary attributes' })
         
         // user not found
         let found = await User.findOne({username: username})
@@ -796,6 +798,9 @@ export const deleteTransaction = async (req, res) => {
         found = await transactions.findOne( {_id: id} )
         if(!found)
             return res.status(400).json({ error: 'transaction not found' })
+
+        if(found.username != username)
+            return res.status(400).json({ error: 'transaction not found for the user' })
 
         const query = { _id: mongoose.Types.ObjectId(req.body.id), username: username }
         const data = await transactions.deleteOne(query);
